refactor: create getSavedTransactions utility function
