+++ src/features/transactionTable/pages/TransactionTablePage.tsx
@@ -3,33 +3,16 @@ import { DataTable } from '@/components/DataTable';
 import { Plus as PlusIcon } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { columns } from '../components/columns/columns';
-import allTransactionJSON from '@/features/transactionTable/mockData/allTransactions.json';
-import { TransactionTableData } from '@/components/AddTransactionForm/AddTransactionForm';
-import { transactionTableDataSchema } from '@/components/AddTransactionForm/transactionTableDataSchema';
 import { useTransactionStore } from '@/stores/TransactionStore';
+import { getSavedTransactions } from '@/utils/getSavedTransactions';
 
 export const TransactionTablePage = () => {
   const defaultSorting = { id: 'transactionDate', desc: true };
 
-  const savedTransactions = allTransactionJSON
-    .map((transaction) => {
-      const parsedTransactions =
-        transactionTableDataSchema.safeParse(transaction);
-      if (parsedTransactions.success) {
-        return parsedTransactions.data;
-      } else {
-        console.error(parsedTransactions.error);
-        return null;
-      }
-    })
-    .filter(
-      (transaction): transaction is TransactionTableData => !!transaction,
-    );
-
+  const savedTransactions = getSavedTransactions();
   const existingTransactions = useTransactionStore(
     (state) => state.transactions,
   );
-
   const transactions = [...existingTransactions, ...savedTransactions];
 
   return (
